LayoutTests/fast/dom/Window/window-function-name-getter-precedence-expected.txt

ALERT: SUCCESS: If you see this alert the test has succeeded.
If you don't see an alert dialog (or ALERT text output if you running this under DRT) this test has failed.
